United trigger_execution config and pandas backend
